Move ClientBenefitsCheck entity logic from client to api
