+++ client/src/Entity/Report/ClientBenefitsCheck.php
@@ -92,14 +92,6 @@ class ClientBenefitsCheck implements ClientBenefitsCheckInterface
     {
         $this->whenLastCheckedEntitlement = $whenLastCheckedEntitlement;
 
-        if (self::WHEN_CHECKED_IVE_NEVER_CHECKED !== $whenLastCheckedEntitlement) {
-            $this->setNeverCheckedExplanation(null);
-        }
-
-        if (self::WHEN_CHECKED_I_HAVE_CHECKED !== $whenLastCheckedEntitlement) {
-            $this->setDateLastCheckedEntitlement(null);
-        }
-
         return $this;
     }
 
@@ -159,10 +151,6 @@ class ClientBenefitsCheck implements ClientBenefitsCheckInterface
     public function setDoOthersReceiveIncomeOnClientsBehalf(?string $doOthersReceiveIncomeOnClientsBehalf): ClientBenefitsCheck
     {
         $this->doOthersReceiveIncomeOnClientsBehalf = $doOthersReceiveIncomeOnClientsBehalf;
-
-        if (self::OTHER_INCOME_DONT_KNOW !== $doOthersReceiveIncomeOnClientsBehalf) {
-            $this->setDontKnowIncomeExplanation(null);
-        }
 
         return $this;
     }
