+++ client/tests/phpunit/Entity/Report/ClientBenefitsCheckTest.php
@@ -90,84 +90,4 @@ class ClientBenefitsCheckTest extends TestCase
             ],
         ];
     }
-
-    /**
-     * @test
-     * @dataProvider whenCheckedProvider
-     */
-    public function setWhenLastCheckedEntitlementSetsNeverCheckedNullWhenValueIsNotNeverChecked(
-        string $whenChecked,
-        ?string $expectedExplanation
-    ) {
-        $sut = (new ClientBenefitsCheck())
-            ->setWhenLastCheckedEntitlement(ClientBenefitsCheck::WHEN_CHECKED_IVE_NEVER_CHECKED)
-            ->setNeverCheckedExplanation('An explanation');
-
-        $sut->setWhenLastCheckedEntitlement($whenChecked);
-
-        self::assertEquals($expectedExplanation, $sut->getNeverCheckedExplanation());
-    }
-
-    public function whenCheckedProvider()
-    {
-        return [
-            'haveChecked' => [ClientBenefitsCheck::WHEN_CHECKED_I_HAVE_CHECKED, null],
-            'currentlyChecking' => [ClientBenefitsCheck::WHEN_CHECKED_IM_CURRENTLY_CHECKING, null],
-            'neverChecked' => [ClientBenefitsCheck::WHEN_CHECKED_IVE_NEVER_CHECKED, 'An explanation'],
-        ];
-    }
-
-    /**
-     * @test
-     * @dataProvider otherIncomeProvider
-     */
-    public function setDoOthersReceiveIncomeOnClientsBehalfSetsDontKnowNullWhenValueIsNotDontKnow(
-        string $otherIncomeReceived,
-        ?string $expectedExplanation
-    ) {
-        $sut = (new ClientBenefitsCheck())
-            ->setDoOthersReceiveIncomeOnClientsBehalf(ClientBenefitsCheck::OTHER_INCOME_DONT_KNOW)
-            ->setDontKnowIncomeExplanation('Another explanation');
-
-        $sut->setDoOthersReceiveIncomeOnClientsBehalf($otherIncomeReceived);
-
-        self::assertEquals($expectedExplanation, $sut->getDontKnowIncomeExplanation());
-    }
-
-    public function otherIncomeProvider()
-    {
-        return [
-            'no' => [ClientBenefitsCheck::OTHER_INCOME_NO, null],
-            'yes' => [ClientBenefitsCheck::OTHER_INCOME_YES, null],
-            'dontKnow' => [ClientBenefitsCheck::OTHER_INCOME_DONT_KNOW, 'Another explanation'],
-        ];
-    }
-
-    /**
-     * @test
-     * @dataProvider dateCheckedProvider
-     */
-    public function setWhenLastCheckedEntitlementSetsDateNullWhenValueIsNotHaveChecked(
-        string $whenChecked,
-        ClientBenefitsCheck $sut,
-        ?DateTime $expectedDateLastChecked
-    ) {
-        $sut->setWhenLastCheckedEntitlement($whenChecked);
-
-        self::assertEquals($expectedDateLastChecked, $sut->getDateLastCheckedEntitlement());
-    }
-
-    public function dateCheckedProvider()
-    {
-        $now = new DateTime();
-        $sut = (new ClientBenefitsCheck())
-            ->setWhenLastCheckedEntitlement(ClientBenefitsCheck::WHEN_CHECKED_I_HAVE_CHECKED)
-            ->setDateLastCheckedEntitlement($now);
-
-        return [
-            'haveChecked' => [ClientBenefitsCheck::WHEN_CHECKED_I_HAVE_CHECKED, $sut, $now],
-            'currentlyChecking' => [ClientBenefitsCheck::WHEN_CHECKED_IM_CURRENTLY_CHECKING, $sut, null],
-            'neverChecked' => [ClientBenefitsCheck::WHEN_CHECKED_IVE_NEVER_CHECKED, $sut, null],
-        ];
-    }
 }
